+++ src/test/java/com/ecommerce/runners/TestRunner.java
@@ -8,7 +8,7 @@ import io.cucumber.junit.CucumberOptions;
 @RunWith(Cucumber.class)
 @CucumberOptions(features = "src/test/java/com/ecommrece/features", 
                  glue = {"com.ecommerce.stepDefinitions", "com.ecommerce.hooks"},
-                 tags = "@test2",
+                 tags = "@test4",
 
                  plugin = {"pretty", "html:target/cucumber-reports.html",
                 		 "com.aventstack.extentreports.cucumber.adapter.ExtentCucumberAdapter:"},
